flipkart/config.py
from dotenv import load_dotenv
import os


load_dotenv()

class Config:
    ASTRA_DB_API_ENDPOINT = os.getenv('ASTRA_DB_ENDPOINT')
    ASTRA_DB_APPLICATION_TOKEN = os.getenv('ASTRA_DB_APPLICTION_TOKEN')
    ASTRA_DB_KEYSPACE= os.getenv('ASTRA_DB_KEYSPACE')
    GROQ_API_KEY = os.getenv('GROQ_API_KEY')
    EMBEDDING_MODEL = "BAAI/bge-base-en-v1.5"
    RAG_MODEL = 'llama-3.1-8b-instant'
